fix: public controller 분리 및 각 api에 대한 swagger 문서 연동 오류 수정
